Project restructure - removed duplicate folder names
